WIP - updating qa-test-app
