+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/base/DrawerActivity.kt
@@ -2,9 +2,14 @@ package com.mapbox.navigation.qa_test_app.view.base
 
 import android.os.Bundle
 import android.view.View
+import android.widget.AdapterView
+import android.widget.SpinnerAdapter
 import androidx.annotation.CallSuper
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.AppCompatSpinner
+import androidx.appcompat.widget.SwitchCompat
 import androidx.core.view.GravityCompat
+import androidx.lifecycle.MutableLiveData
 import com.mapbox.navigation.qa_test_app.databinding.LayoutActivityDrawerActivityBinding
 
 abstract class DrawerActivity : AppCompatActivity() {
@@ -32,5 +37,62 @@ abstract class DrawerActivity : AppCompatActivity() {
 
     fun closeDrawers() {
         binding.drawerLayout.closeDrawers()
+    }
+
+    protected fun bindSwitch(
+        switch: SwitchCompat,
+        getValue: () -> Boolean,
+        setValue: (v: Boolean) -> Unit
+    ) {
+        switch.isChecked = getValue()
+        switch.setOnCheckedChangeListener { _, isChecked -> setValue(isChecked) }
+    }
+
+    protected fun bindSwitch(
+        switch: SwitchCompat,
+        liveData: MutableLiveData<Boolean>,
+        onChange: (value: Boolean) -> Unit
+    ) {
+        liveData.observe(this) {
+            switch.isChecked = it
+            onChange(it)
+        }
+        switch.setOnCheckedChangeListener { _, isChecked ->
+            liveData.value = isChecked
+        }
+    }
+
+    protected fun bindSpinner(
+        spinner: AppCompatSpinner,
+        liveData: MutableLiveData<String>,
+        onChange: (value: String) -> Unit
+    ) {
+        liveData.observe(this) {
+            if (spinner.selectedItem != it) {
+                spinner.setSelection(spinner.adapter.findItemPosition(it) ?: 0)
+            }
+            onChange(it)
+        }
+
+        spinner.onItemSelectedListener =
+            object : AdapterView.OnItemSelectedListener {
+                override fun onItemSelected(
+                    parent: AdapterView<*>,
+                    view: View?,
+                    position: Int,
+                    id: Long
+                ) {
+                    liveData.value = parent.getItemAtPosition(position) as? String
+                }
+
+                override fun onNothingSelected(parent: AdapterView<*>?) = Unit
+            }
+    }
+
+    private fun SpinnerAdapter.findItemPosition(item: Any): Int? {
+        for (pos in 0..count) {
+            if (item == getItem(pos)) return pos
+        }
+        return null
     }
 }
