+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/customnavview/MapboxNavigationViewCustomizedActivity.kt
@@ -8,20 +8,15 @@ import android.view.Gravity
 import android.view.View
 import android.view.ViewGroup
 import android.view.ViewGroup.LayoutParams.WRAP_CONTENT
-import android.widget.AdapterView
 import android.widget.LinearLayout.LayoutParams
-import android.widget.SpinnerAdapter
 import android.widget.Toast
 import androidx.activity.addCallback
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatDelegate
-import androidx.appcompat.widget.AppCompatSpinner
 import androidx.appcompat.widget.AppCompatTextView
-import androidx.appcompat.widget.SwitchCompat
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.toBitmap
 import androidx.core.view.WindowCompat
-import androidx.lifecycle.MutableLiveData
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.mapbox.api.directions.v5.DirectionsCriteria
 import com.mapbox.api.directions.v5.models.RouteOptions
@@ -88,17 +83,7 @@ import com.mapbox.navigation.qa_test_app.dump.NavigationViewApiDumpInterceptor
 import com.mapbox.navigation.qa_test_app.view.base.DrawerActivity
 import com.mapbox.navigation.ui.base.lifecycle.UIBinder
 import com.mapbox.navigation.ui.base.lifecycle.UIComponent
-import com.mapbox.navigation.ui.maneuver.model.ManeuverExitOptions
-import com.mapbox.navigation.ui.maneuver.model.ManeuverPrimaryOptions
-import com.mapbox.navigation.ui.maneuver.model.ManeuverSecondaryOptions
-import com.mapbox.navigation.ui.maneuver.model.ManeuverSubOptions
-import com.mapbox.navigation.ui.maneuver.model.ManeuverViewOptions
 import com.mapbox.navigation.ui.maps.NavigationStyles
-import com.mapbox.navigation.ui.maps.route.RouteLayerConstants
-import com.mapbox.navigation.ui.maps.route.arrow.model.RouteArrowOptions
-import com.mapbox.navigation.ui.maps.route.line.model.MapboxRouteLineOptions
-import com.mapbox.navigation.ui.maps.route.line.model.RouteLineColorResources
-import com.mapbox.navigation.ui.maps.route.line.model.RouteLineResources
 import com.mapbox.navigation.utils.internal.toPoint
 
 @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
@@ -106,52 +91,12 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
 
     private val viewModel: CustomizedViewModel by viewModels()
 
-    private val navViewListener =
-        LoggingNavigationViewListener("MapboxNavigationViewCustomizedActivity")
-
-    private val onMapLongClick = object : MapViewObserver(), OnMapLongClickListener {
-
-        override fun onAttached(mapView: MapView) {
-            mapView.gestures.addOnMapLongClickListener(this)
-        }
-
-        override fun onDetached(mapView: MapView) {
-            mapView.gestures.removeOnMapLongClickListener(this)
-        }
-
-        override fun onMapLongClick(point: Point): Boolean {
-            val lastLocation = lastLocation ?: return false
-            val mapboxNavigation = MapboxNavigationApp.current() ?: return false
-            val routeOptions = RouteOptions.builder()
-                .applyDefaultNavigationOptions()
-                .applyLanguageAndVoiceUnitOptions(this@MapboxNavigationViewCustomizedActivity)
-                .coordinatesList(listOf(lastLocation.toPoint(), point))
-                .build()
-            val callback = object : NavigationRouterCallback {
-
-                override fun onRoutesReady(
-                    routes: List<NavigationRoute>,
-                    routerOrigin: RouterOrigin,
-                ) {
-                    binding.navigationView.api.startRoutePreview(routes)
-                }
-
-                override fun onFailure(reasons: List<RouterFailure>, routeOptions: RouteOptions) {
-                    // no impl
-                }
-
-                override fun onCanceled(routeOptions: RouteOptions, routerOrigin: RouterOrigin) {
-                    // no impl
-                }
-            }
-
-            mapboxNavigation.requestRoutes(routeOptions, callback)
-            return false
-        }
-    }
-
     private lateinit var binding: LayoutActivityNavigationViewBinding
     private lateinit var menuBinding: LayoutDrawerMenuNavViewCustomBinding
+
+    private var lastLocation: Location? = null
+    private lateinit var customInfoPanel: CustomInfoPanelBinder
+    private lateinit var customInfoPanelFixedHeight: CustomInfoPanelBinderWithFixedHeight
 
     override fun onCreateContentView(): View {
         binding = LayoutActivityNavigationViewBinding.inflate(layoutInflater)
@@ -163,57 +108,11 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         return menuBinding.root
     }
 
-    private fun updateTheme() {
-        if (viewModel.fullScreen.value == true) {
-            setTheme(R.style.Theme_Fullscreen)
-            WindowCompat.setDecorFitsSystemWindows(window, false)
-        } else {
-            setTheme(R.style.Theme_AppCompat_NoActionBar)
-            WindowCompat.setDecorFitsSystemWindows(window, true)
-        }
-    }
+    @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        updateTheme()
+        super.onCreate(savedInstanceState)
 
-    /**
-     * This is a feature for Mapbox Navigation development. When the notification service is
-     * available you are able to send commands and handle state changes with dumpsys.
-     *
-     * $ adb shell dumpsys activity service com.mapbox.navigation.core.trip.service.NavigationNotificationService help
-     */
-    private val dumpCommands = object : MapboxNavigationObserver {
-
-        private val interceptors by lazy {
-            arrayOf(
-                DistanceFormatterDumpInterceptor(),
-                NavigationViewApiDumpInterceptor(binding.navigationView.api)
-                // Add your interceptors here
-            )
-        }
-
-        override fun onAttached(mapboxNavigation: MapboxNavigation) {
-            MapboxDumpRegistry.addInterceptors(*interceptors)
-        }
-
-        override fun onDetached(mapboxNavigation: MapboxNavigation) {
-            MapboxDumpRegistry.removeInterceptors(*interceptors)
-        }
-    }
-
-    private val locationObserver = object : LocationObserver {
-
-        override fun onNewRawLocation(rawLocation: Location) {
-            // no impl
-        }
-
-        override fun onNewLocationMatcherResult(locationMatcherResult: LocationMatcherResult) {
-            lastLocation = locationMatcherResult.enhancedLocation
-        }
-    }
-
-    private var lastLocation: Location? = null
-    private lateinit var customInfoPanel: CustomInfoPanelBinder
-    private lateinit var customInfoPanelFixedHeight: CustomInfoPanelBinderWithFixedHeight
-
-    init {
         attachResumed(dumpCommands)
         attachCreated(
             object : MapboxNavigationObserver {
@@ -227,12 +126,6 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
                 }
             },
         )
-    }
-
-    @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
-    override fun onCreate(savedInstanceState: Bundle?) {
-        updateTheme()
-        super.onCreate(savedInstanceState)
 
         customInfoPanel = CustomInfoPanelBinder(binding.navigationView)
         customInfoPanelFixedHeight = CustomInfoPanelBinderWithFixedHeight(binding.navigationView)
@@ -241,118 +134,35 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         binding.navigationView.addListener(navViewListener)
 
         AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.getDefaultNightMode())
-
-        bindSwitch(
-            menuBinding.toggleViewStyling,
-            viewModel.useCustomStyles,
-            ::customizeStyles
-        )
-
-        bindSwitch(
-            menuBinding.toggleCustomViews,
-            viewModel.useCustomViews,
-            ::customizeViews
-        )
-
-        bindSwitch(
-            menuBinding.toggleCustomMap,
-            viewModel.showCustomMapView,
-            ::customizeMap
-        )
-
-        bindSwitch(
-            menuBinding.toggleFullscreen,
-            getValue = { viewModel.fullScreen.value == true },
-            setValue = {
-                viewModel.fullScreen.value = it
-                recreate()
-            }
-        )
-
-        bindSpinner(
-            menuBinding.spinnerCustomInfoPanelContent,
-            viewModel.customInfoPanelContent,
-            ::setCustomInfoPanelContent
-        )
-        bindSpinner(
-            menuBinding.spinnerCustomInfoPanelPeekHeight,
-            viewModel.customInfoPanelPeekHeight,
-            ::setCustomInfoPanelPeekHeight,
-        )
-        bindSwitch(
-            menuBinding.toggleCustomInfoPanelEndNavButton,
-            viewModel.showCustomInfoPanelEndNavButton,
-            ::toggleCustomInfoPanelEndNavButton
-        )
-        bindSpinner(
-            menuBinding.spinnerCustomInfoPanelLayout,
-            viewModel.customInfoPanelLayout,
-            ::setCustomInfoPanelLayout
-        )
-        bindSwitch(
-            menuBinding.toggleCustomInfoPanelStyles,
-            viewModel.useCustomInfoPanelStyles,
-            ::toggleCustomInfoPanelStyles
-        )
-
-        bindSwitch(
-            menuBinding.toggleBottomSheetFD,
-            viewModel.showBottomSheetInFreeDrive,
-            ::toggleShowInfoPanelInFreeDrive
-        )
-
-        bindSwitch(
-            menuBinding.toggleEnableDestinationPreview,
-            viewModel.enableDestinationPreview,
-            ::toggleEnableDestinationPreview,
-        )
-
-        bindSwitch(
-            menuBinding.toggleIsInfoPanelHideable,
-            viewModel.isInfoPanelHideable,
-            ::toggleInfoPanelHiding
-        )
-
-        bindSwitch(
-            menuBinding.useMetric,
-            viewModel.distanceFormatterMetric,
-            ::toggleUseMetric
-        )
-
-        bindSpinner(
-            menuBinding.spinnerInfoPanelVisibilityOverride,
-            viewModel.infoPanelStateOverride,
-            ::overrideInfoPanelState
-        )
-
-        bindSwitch(
-            menuBinding.showCameraDebugInfo,
-            viewModel.showCameraDebugInfo,
-            ::toggleShowCameraDebugInfo
-        )
-
-        bindSwitch(
-            menuBinding.toggleEnableScalebar,
-            viewModel.enableScalebar,
-            ::toggleEnableScalebar
-        )
-
-        bindSpinner(
-            menuBinding.spinnerProfile,
-            viewModel.routingProfile,
-            ::setRoutingProfile,
-        )
-
-        bindSwitch(
-            menuBinding.toggleEnableCompass,
-            viewModel.enableCompass,
-            ::toggleEnableCompass
-        )
     }
 
     override fun onResume() {
         super.onResume()
 
+        bindSwitch(
+            menuBinding.toggleReplay,
+            getValue = binding.navigationView.api::isReplayEnabled,
+            setValue = { isChecked ->
+                binding.navigationView.api.routeReplayEnabled(isChecked)
+            }
+        )
+        initActivityOptions()
+        initGeneralOptions()
+        initMapOptions()
+        initInfoPanelOptions()
+    }
+
+    private fun updateTheme() {
+        if (viewModel.fullScreen.value == true) {
+            setTheme(R.style.Theme_Fullscreen)
+            WindowCompat.setDecorFitsSystemWindows(window, false)
+        } else {
+            setTheme(R.style.Theme_AppCompat_NoActionBar)
+            WindowCompat.setDecorFitsSystemWindows(window, true)
+        }
+    }
+
+    private fun initActivityOptions() {
         bindSwitch(
             menuBinding.toggleTheme,
             getValue = resources.configuration::isNightMode,
@@ -368,79 +178,47 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         )
 
         bindSwitch(
-            menuBinding.toggleReplay,
-            getValue = binding.navigationView.api::isReplayEnabled,
-            setValue = { isChecked ->
-                binding.navigationView.api.routeReplayEnabled(isChecked)
+            menuBinding.toggleFullscreen,
+            getValue = { viewModel.fullScreen.value == true },
+            setValue = {
+                viewModel.fullScreen.value = it
+                recreate()
             }
         )
     }
 
-    private fun bindSwitch(
-        switch: SwitchCompat,
-        getValue: () -> Boolean,
-        setValue: (v: Boolean) -> Unit
-    ) {
-        switch.isChecked = getValue()
-        switch.setOnCheckedChangeListener { _, isChecked -> setValue(isChecked) }
+    //region General Options
+
+    private fun initGeneralOptions() {
+        bindSpinner(
+            menuBinding.spinnerProfile,
+            viewModel.routingProfile,
+            ::toggleRoutingProfile,
+        )
+        bindSwitch(
+            menuBinding.toggleViewStyling,
+            viewModel.useCustomStyles,
+            ::toggleCustomStyles
+        )
+        bindSwitch(
+            menuBinding.toggleCustomViews,
+            viewModel.useCustomViews,
+            ::toggleCustomViews
+        )
+        bindSwitch(
+            menuBinding.useMetric,
+            viewModel.distanceFormatterMetric,
+            ::toggleUseMetric
+        )
+        bindSwitch(
+            menuBinding.showCameraDebugInfo,
+            viewModel.showCameraDebugInfo,
+            ::toggleShowCameraDebugInfo
+        )
     }
 
-    private fun bindSwitch(
-        switch: SwitchCompat,
-        liveData: MutableLiveData<Boolean>,
-        onChange: (value: Boolean) -> Unit
-    ) {
-        liveData.observe(this) {
-            switch.isChecked = it
-            onChange(it)
-        }
-        switch.setOnCheckedChangeListener { _, isChecked ->
-            liveData.value = isChecked
-        }
-    }
-
-    private fun bindSpinner(
-        spinner: AppCompatSpinner,
-        liveData: MutableLiveData<String>,
-        onChange: (value: String) -> Unit
-    ) {
-        liveData.observe(this) {
-            if (spinner.selectedItem != it) {
-                spinner.setSelection(spinner.adapter.findItemPosition(it) ?: 0)
-            }
-            onChange(it)
-        }
-
-        spinner.onItemSelectedListener =
-            object : AdapterView.OnItemSelectedListener {
-                override fun onItemSelected(
-                    parent: AdapterView<*>,
-                    view: View?,
-                    position: Int,
-                    id: Long
-                ) {
-                    liveData.value = parent.getItemAtPosition(position) as? String
-                }
-
-                override fun onNothingSelected(parent: AdapterView<*>?) = Unit
-            }
-    }
-
-    private fun overrideInfoPanelState(value: String) {
-        val newState = when (value) {
-            "HIDDEN" -> BottomSheetBehavior.STATE_HIDDEN
-            "COLLAPSED" -> BottomSheetBehavior.STATE_COLLAPSED
-            "HALF_EXPANDED" -> BottomSheetBehavior.STATE_HALF_EXPANDED
-            "EXPANDED" -> BottomSheetBehavior.STATE_EXPANDED
-            else -> 0
-        }
-        binding.navigationView.customizeViewOptions {
-            infoPanelForcedState = newState
-        }
-    }
-
-    private fun customizeStyles(customStyles: Boolean) {
-        if (customStyles) {
+    private fun toggleCustomStyles(enabled: Boolean) {
+        if (enabled) {
             binding.navigationView.customizeViewStyles {
                 tripProgressStyle = R.style.MyCustomTripProgressStyle
                 speedLimitStyle = R.style.MyCustomSpeedLimitStyle
@@ -517,10 +295,10 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
     }
 
-    private fun customizeViews(showCustomViews: Boolean) {
+    private fun toggleCustomViews(enabled: Boolean) {
         // This demonstrates that you can customize views at any time. You can also reset to
         // the default views.
-        if (showCustomViews) {
+        if (enabled) {
             binding.navigationView.customizeViewBinders {
                 speedLimitBinder = CustomSpeedLimitViewBinder()
                 customActionButtons = listOf(
@@ -531,8 +309,8 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
                 )
             }
             binding.navigationView.customizeViewOptions {
-                routeLineOptions = customRouteLineOptions()
-                routeArrowOptions = customRouteArrowOptions()
+                routeLineOptions = customRouteLineOptions(applicationContext)
+                routeArrowOptions = customRouteArrowOptions(applicationContext)
                 mapStyleUriDay = Style.LIGHT
                 mapStyleUriNight = Style.DARK
             }
@@ -557,9 +335,70 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
     }
 
-    private fun customizeMap(showCustomMapView: Boolean) {
+    private fun toggleUseMetric(showMetric: Boolean) {
+        val options = DistanceFormatterOptions
+            .Builder(this@MapboxNavigationViewCustomizedActivity)
+            .build()
+        if (showMetric) {
+            binding.navigationView.customizeViewOptions {
+                distanceFormatterOptions = options.toBuilder().unitType(UnitType.METRIC).build()
+            }
+        } else {
+            binding.navigationView.customizeViewOptions {
+                distanceFormatterOptions = options.toBuilder().unitType(UnitType.IMPERIAL).build()
+            }
+        }
+    }
+
+    private fun toggleRoutingProfile(profile: String) {
+        val routingProfile = when (profile) {
+            "DRIVING-TRAFFIC" -> DirectionsCriteria.PROFILE_DRIVING_TRAFFIC
+            "DRIVING" -> DirectionsCriteria.PROFILE_DRIVING
+            "WALKING" -> DirectionsCriteria.PROFILE_WALKING
+            "CYCLING" -> DirectionsCriteria.PROFILE_CYCLING
+            else -> {
+                binding.navigationView.setRouteOptionsInterceptor(null)
+                return
+            }
+        }
+        binding.navigationView.setRouteOptionsInterceptor { defaultBuilder ->
+            defaultBuilder.layers(null).applyDefaultNavigationOptions(routingProfile)
+        }
+    }
+
+    private fun toggleShowCameraDebugInfo(enabled: Boolean) {
+        binding.navigationView.customizeViewOptions {
+            showCameraDebugInfo = enabled
+        }
+    }
+
+    //endregion
+
+    //region Map Options
+
+    private fun initMapOptions() {
+        bindSwitch(
+            menuBinding.toggleCustomMap,
+            viewModel.showCustomMapView,
+            ::toggleCustomMap
+        )
+
+        bindSwitch(
+            menuBinding.toggleEnableScalebar,
+            viewModel.enableScalebar,
+            ::toggleEnableScalebar
+        )
+
+        bindSwitch(
+            menuBinding.toggleEnableCompass,
+            viewModel.enableCompass,
+            ::toggleEnableCompass
+        )
+    }
+
+    private fun toggleCustomMap(enabled: Boolean) {
         // Demonstrate map customization
-        if (showCustomMapView) {
+        if (enabled) {
             binding.navigationView.customizeViewBinders {
                 mapViewBinder = CustomMapViewBinder()
             }
@@ -570,7 +409,99 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
     }
 
-    private fun setCustomInfoPanelContent(contentSize: String) {
+    private fun toggleEnableScalebar(enabled: Boolean) {
+        binding.navigationView.customizeViewStyles {
+            mapScalebarParams = MapboxMapScalebarParams
+                .Builder(this@MapboxNavigationViewCustomizedActivity)
+                .enabled(enabled)
+                .build()
+        }
+    }
+
+    private fun toggleEnableCompass(enabled: Boolean) {
+        binding.navigationView.customizeViewStyles {
+            compassButtonParams = ViewStyleCustomization.defaultCompassButtonParams(
+                this@MapboxNavigationViewCustomizedActivity
+            ).let {
+                MapboxExtendableButtonParams(
+                    it.style,
+                    it.layoutParams,
+                    enabled
+                )
+            }
+        }
+    }
+
+    //endregion
+
+    //region Info Panel Options
+
+    private fun initInfoPanelOptions() {
+        bindSpinner(
+            menuBinding.spinnerCustomInfoPanelContent,
+            viewModel.customInfoPanelContent,
+            ::toggleCustomInfoPanelContent
+        )
+        bindSpinner(
+            menuBinding.spinnerCustomInfoPanelPeekHeight,
+            viewModel.customInfoPanelPeekHeight,
+            ::toggleCustomInfoPanelPeekHeight,
+        )
+        bindSwitch(
+            menuBinding.toggleCustomInfoPanelEndNavButton,
+            viewModel.showCustomInfoPanelEndNavButton,
+            ::toggleCustomInfoPanelEndNavButton
+        )
+        bindSpinner(
+            menuBinding.spinnerCustomInfoPanelLayout,
+            viewModel.customInfoPanelLayout,
+            ::toggleCustomInfoPanelLayout
+        )
+        bindSwitch(
+            menuBinding.toggleCustomInfoPanelStyles,
+            viewModel.useCustomInfoPanelStyles,
+            ::toggleCustomInfoPanelStyles
+        )
+
+        bindSwitch(
+            menuBinding.toggleBottomSheetFD,
+            viewModel.showBottomSheetInFreeDrive,
+            ::toggleShowInfoPanelInFreeDrive
+        )
+
+        bindSwitch(
+            menuBinding.toggleEnableDestinationPreview,
+            viewModel.enableDestinationPreview,
+            ::toggleEnableDestinationPreview,
+        )
+
+        bindSwitch(
+            menuBinding.toggleIsInfoPanelHideable,
+            viewModel.isInfoPanelHideable,
+            ::toggleInfoPanelHiding
+        )
+
+        bindSpinner(
+            menuBinding.spinnerInfoPanelVisibilityOverride,
+            viewModel.infoPanelStateOverride,
+            ::toggleInfoPanelState
+        )
+    }
+
+    private fun toggleInfoPanelState(state: String) {
+        val newState = when (state) {
+            "HIDDEN" -> BottomSheetBehavior.STATE_HIDDEN
+            "COLLAPSED" -> BottomSheetBehavior.STATE_COLLAPSED
+            "HALF_EXPANDED" -> BottomSheetBehavior.STATE_HALF_EXPANDED
+            "EXPANDED" -> BottomSheetBehavior.STATE_EXPANDED
+            else -> 0
+        }
+        binding.navigationView.customizeViewOptions {
+            infoPanelForcedState = newState
+        }
+    }
+
+    private fun toggleCustomInfoPanelContent(contentSize: String) {
         binding.navigationView.customizeViewBinders {
             when (contentSize) {
                 "SMALL",
@@ -592,7 +523,7 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
     }
 
-    private fun setCustomInfoPanelPeekHeight(peekHeight: String) {
+    private fun toggleCustomInfoPanelPeekHeight(peekHeight: String) {
         val defaultPeekHeight = ViewStyleCustomization.defaultInfoPanelPeekHeight(context = this)
         binding.navigationView.customizeViewStyles {
             infoPanelPeekHeight = when (peekHeight) {
@@ -612,7 +543,7 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
     }
 
-    private fun setCustomInfoPanelLayout(selection: String) {
+    private fun toggleCustomInfoPanelLayout(selection: String) {
         when (selection) {
             "CUSTOM" -> {
                 customInfoPanel.setEnabled(true)
@@ -676,59 +607,7 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
     }
 
-    private fun toggleUseMetric(showMetric: Boolean) {
-        val options = DistanceFormatterOptions
-            .Builder(this@MapboxNavigationViewCustomizedActivity)
-            .build()
-        if (showMetric) {
-            binding.navigationView.customizeViewOptions {
-                distanceFormatterOptions = options.toBuilder().unitType(UnitType.METRIC).build()
-            }
-        } else {
-            binding.navigationView.customizeViewOptions {
-                distanceFormatterOptions = options.toBuilder().unitType(UnitType.IMPERIAL).build()
-            }
-        }
-    }
-
-    private fun toggleEnableScalebar(enabled: Boolean) {
-        binding.navigationView.customizeViewStyles {
-            mapScalebarParams = MapboxMapScalebarParams
-                .Builder(this@MapboxNavigationViewCustomizedActivity)
-                .enabled(enabled)
-                .build()
-        }
-    }
-
-    private fun setRoutingProfile(profile: String) {
-        val routingProfile = when (profile) {
-            "DRIVING-TRAFFIC" -> DirectionsCriteria.PROFILE_DRIVING_TRAFFIC
-            "DRIVING" -> DirectionsCriteria.PROFILE_DRIVING
-            "WALKING" -> DirectionsCriteria.PROFILE_WALKING
-            "CYCLING" -> DirectionsCriteria.PROFILE_CYCLING
-            else -> {
-                binding.navigationView.setRouteOptionsInterceptor(null)
-                return
-            }
-        }
-        binding.navigationView.setRouteOptionsInterceptor { defaultBuilder ->
-            defaultBuilder.layers(null).applyDefaultNavigationOptions(routingProfile)
-        }
-    }
-
-    private fun toggleEnableCompass(enabled: Boolean) {
-        binding.navigationView.customizeViewStyles {
-            compassButtonParams = ViewStyleCustomization.defaultCompassButtonParams(
-                this@MapboxNavigationViewCustomizedActivity
-            ).let {
-                MapboxExtendableButtonParams(
-                    it.style,
-                    it.layoutParams,
-                    enabled
-                )
-            }
-        }
-    }
+    //endregion
 
     private fun customActionButton(text: String): View {
         return AppCompatTextView(this).apply {
@@ -745,78 +624,83 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
     }
 
-    private fun customRouteLineOptions() =
-        MapboxRouteLineOptions.Builder(this)
-            .withRouteLineResources(
-                RouteLineResources.Builder()
-                    .routeLineColorResources(
-                        RouteLineColorResources.Builder()
-                            .routeLowCongestionColor(Color.YELLOW)
-                            .routeCasingColor(Color.RED)
-                            .build()
-                    )
-                    .build()
+    private val navViewListener =
+        LoggingNavigationViewListener("MapboxNavigationViewCustomizedActivity")
+
+    private val onMapLongClick = object : MapViewObserver(), OnMapLongClickListener {
+
+        override fun onAttached(mapView: MapView) {
+            mapView.gestures.addOnMapLongClickListener(this)
+        }
+
+        override fun onDetached(mapView: MapView) {
+            mapView.gestures.removeOnMapLongClickListener(this)
+        }
+
+        override fun onMapLongClick(point: Point): Boolean {
+            val lastLocation = lastLocation ?: return false
+            val mapboxNavigation = MapboxNavigationApp.current() ?: return false
+            val routeOptions = RouteOptions.builder()
+                .applyDefaultNavigationOptions()
+                .applyLanguageAndVoiceUnitOptions(this@MapboxNavigationViewCustomizedActivity)
+                .coordinatesList(listOf(lastLocation.toPoint(), point))
+                .build()
+            val callback = object : NavigationRouterCallback {
+
+                override fun onRoutesReady(
+                    routes: List<NavigationRoute>,
+                    routerOrigin: RouterOrigin,
+                ) {
+                    binding.navigationView.api.startRoutePreview(routes)
+                }
+
+                override fun onFailure(reasons: List<RouterFailure>, routeOptions: RouteOptions) {
+                    // no impl
+                }
+
+                override fun onCanceled(routeOptions: RouteOptions, routerOrigin: RouterOrigin) {
+                    // no impl
+                }
+            }
+
+            mapboxNavigation.requestRoutes(routeOptions, callback)
+            return false
+        }
+    }
+
+    /**
+     * This is a feature for Mapbox Navigation development. When the notification service is
+     * available you are able to send commands and handle state changes with dumpsys.
+     *
+     * $ adb shell dumpsys activity service com.mapbox.navigation.core.trip.service.NavigationNotificationService help
+     */
+    private val dumpCommands = object : MapboxNavigationObserver {
+
+        private val interceptors by lazy {
+            arrayOf(
+                DistanceFormatterDumpInterceptor(),
+                NavigationViewApiDumpInterceptor(binding.navigationView.api)
+                // Add your interceptors here
             )
-            .withRouteLineBelowLayerId("road-label") // for Style.LIGHT and Style.DARK
-            .withVanishingRouteLineEnabled(true)
-            .displaySoftGradientForTraffic(true)
-            .build()
+        }
 
-    private fun customRouteArrowOptions() =
-        RouteArrowOptions.Builder(this)
-            .withAboveLayerId(RouteLayerConstants.TOP_LEVEL_ROUTE_LINE_LAYER_ID)
-            .withArrowColor(Color.RED)
-            .build()
+        override fun onAttached(mapboxNavigation: MapboxNavigation) {
+            MapboxDumpRegistry.addInterceptors(*interceptors)
+        }
 
-    private fun customManeuverOptions() = ManeuverViewOptions
-        .Builder()
-        .maneuverBackgroundColor(R.color.maneuver_main_background)
-        .subManeuverBackgroundColor(R.color.maneuver_sub_background)
-        .upcomingManeuverBackgroundColor(R.color.maneuver_sub_background)
-        .turnIconManeuver(R.style.MyCustomTurnIconManeuver)
-        .laneGuidanceTurnIconManeuver(R.style.MyCustomTurnIconManeuver)
-        .stepDistanceTextAppearance(R.style.MyCustomStepDistance)
-        .primaryManeuverOptions(
-            ManeuverPrimaryOptions
-                .Builder()
-                .textAppearance(R.style.MyCustomPrimaryManeuver)
-                .exitOptions(
-                    ManeuverExitOptions
-                        .Builder()
-                        .textAppearance(R.style.MyCustomExitTextForPrimary)
-                        .build()
-                )
-                .build()
-        )
-        .secondaryManeuverOptions(
-            ManeuverSecondaryOptions
-                .Builder()
-                .textAppearance(R.style.MyCustomSecondaryManeuver)
-                .exitOptions(
-                    ManeuverExitOptions
-                        .Builder()
-                        .textAppearance(R.style.MyCustomExitTextForSecondary)
-                        .build()
-                )
-                .build()
-        )
-        .subManeuverOptions(
-            ManeuverSubOptions
-                .Builder()
-                .textAppearance(R.style.MyCustomSubManeuver)
-                .exitOptions(
-                    ManeuverExitOptions
-                        .Builder()
-                        .textAppearance(R.style.MyCustomExitTextForSub)
-                        .build()
-                )
-                .build()
-        )
-        .build()
+        override fun onDetached(mapboxNavigation: MapboxNavigation) {
+            MapboxDumpRegistry.removeInterceptors(*interceptors)
+        }
+    }
 
-    private fun toggleShowCameraDebugInfo(show: Boolean) {
-        binding.navigationView.customizeViewOptions {
-            showCameraDebugInfo = show
+    private val locationObserver = object : LocationObserver {
+
+        override fun onNewRawLocation(rawLocation: Location) {
+            // no impl
+        }
+
+        override fun onNewLocationMatcherResult(locationMatcherResult: LocationMatcherResult) {
+            lastLocation = locationMatcherResult.enhancedLocation
         }
     }
 
@@ -890,11 +774,4 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         onBackPressedDispatcher.addCallback(owner = this, enabled = false) {
             binding.navigationView.api.startFreeDrive()
         }
-
-    private fun SpinnerAdapter.findItemPosition(item: Any): Int? {
-        for (pos in 0..count) {
-            if (item == getItem(pos)) return pos
-        }
-        return null
-    }
 }
